requirements.txt
pandas==1.3.3
numpy==1.21.2
pyspark
psycopg2-binary
boto3
botocore==1.17.63
pyyaml